gls: Treat parcelshop scan on first day as 'posted' event
Since GLS Parcelshop data is apparently not immediately processed, it can
happen that a parcel is reported as "in Parcelshop" even after it has already
been scanned in a depot. Work around this by considering all "in parcelshop"
events before/on the same day as the first on-location scan as posted event.